dotnet

+++ EntitySystem2/dotnet/engine/Time.cs
@@ -3,8 +3,8 @@ namespace engine;
 public class Time
 {
     private long time;
-    public double asSeconds() {
-        return time / 1000000.0;
+    public float asSeconds() {
+        return (float)(time / 1000000.0);
     }
 
     public Time(long time)
